8588 - Renamed files from deliveries to Daily data (more explicit) - fixed date filter to keep date being searched upon - fixed clear button not working
