+++ frontend-react/src/components/MessageTracker/MessageReceivers.tsx
@@ -4,7 +4,7 @@ import classnames from "classnames";
 import { ReceiverData } from "../../config/endpoints/messageTracker";
 import { parseFileLocation } from "../../utils/misc";
 import { Table } from "../../shared/Table/Table";
-import ReportLink from "../../pages/deliveries/Table/ReportLink";
+import ReportLink from "../../pages/deliveries/daily-data/ReportLink";
 
 const NO_DATA_STRING = "N/A";
