+++ frontend-react/src/pages/deliveries/daily-data/ReportsUtils.ts
